requirements.txt
learn2learn 
torch 
torchvision


argparse
datetime
matplotlib
numpy
pillow
quadprog
torchvision
urllib3
uuid
tqdm
wandb
scipy